feat: implement CRM deal stage management with controller, service, and service implementation
