+++ backend/src/main/java/com/skapp/community/crmplanner/service/CrmDealService.java
@@ -10,6 +10,4 @@ public interface CrmDealService {
 
 	ResponseEntityDto getDeals(CrmDealFilterDto filterDto);
 
-	ResponseEntityDto getDealStages();
-
 }
