+++ backend/src/main/java/com/skapp/community/crmplanner/service/CrmDealStageService.java
@@ -1,0 +1,9 @@
+package com.skapp.community.crmplanner.service;
+
+import com.skapp.community.common.payload.response.ResponseEntityDto;
+
+public interface CrmDealStageService {
+
+	ResponseEntityDto getDealStages();
+
+}
