+++ backend/src/main/java/com/skapp/community/crmplanner/service/impl/CrmDealStageServiceImpl.java
@@ -1,0 +1,34 @@
+package com.skapp.community.crmplanner.service.impl;
+
+import com.skapp.community.common.payload.response.ResponseEntityDto;
+import com.skapp.community.crmplanner.mapper.CrmDealMapper;
+import com.skapp.community.crmplanner.payload.response.CrmDealStageResponseDto;
+import com.skapp.community.crmplanner.repository.CrmDealStageDao;
+import com.skapp.community.crmplanner.service.CrmDealStageService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class CrmDealStageServiceImpl implements CrmDealStageService {
+
+	private final CrmDealStageDao crmDealStageDao;
+
+	private final CrmDealMapper crmDealMapper;
+
+	@Override
+	@Transactional(readOnly = true)
+	public ResponseEntityDto getDealStages() {
+		log.info("getDealStages: execution started");
+		List<CrmDealStageResponseDto> stages = crmDealMapper
+			.crmDealStagesToCrmDealStageResponseDtos(crmDealStageDao.findAllByIsDeletedFalseOrderByOrderIndexAsc());
+		log.info("getDealStages: execution ended with {} result(s)", stages.size());
+		return new ResponseEntityDto(false, stages);
+	}
+
+}
